refactor, consolidate headers
