+++ puffinsystems_homepage/src/lib/components/Header.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+  export let id: string;
+  export let text: string;
+</script>
+
+<h1 class="my-1 text-4xl font-medium" {id}>
+  {text}
+</h1>
