+++ puffinsystems_homepage/src/lib/components/cardCarousel/cardCarousel.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { CardContent } from "$lib/types/card";
   import { onMount } from "svelte";
-  import ProjectCard from "$lib/components/cardTemplates/ProjectCard.svelte";
+  import CarouselCardTemplate from "$lib/components/cardCarousel/carouselCardTemplate.svelte";
   import CarouselButtons from "$lib/components/cardCarousel/carouselButtons.svelte";
 
   export let cardContentList: CardContent[];
@@ -28,7 +28,7 @@
 <div class="carousel w-full" bind:this={carousel} on:scroll={parseScroll}>
   {#each cardContentList as cardContent, i}
     <div id="{carouselName}-{i + 1}" class="carousel-item w-full">
-      <ProjectCard {cardContent} />
+      <CarouselCardTemplate {cardContent} />
     </div>
   {/each}
 </div>
